mbench: Integrate with meson
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
PR: https://github.com/cloudkernels/vaccelrt/pull/109

+++ plugins/mbench/vaccel.c
@@ -40,7 +40,7 @@ static int mbench(int time)
 	while (1) {
 		clock_gettime(CLOCK_MONOTONIC_RAW, &t);
 		ets = t.tv_sec * 1e9 + t.tv_nsec; // nsec
-		if ((ets-sts)/1000000 >= time)
+		if ((ets-sts)/1000000 >= (uint64_t)time)
 			break;
 	}
